+++ src/django/api/tests/base_moderation_events_production_location_test.py
@@ -171,11 +171,11 @@ class BaseModerationEventsProductionLocationTest(APITestCase):
             self.superuser.id
         )
 
-    def assert_source_creation(self, source):
+    def assert_source_creation(self, source, is_active=True, is_public=True):
         self.assertIsNotNone(source)
         self.assertEqual(source.source_type, Source.SINGLE)
-        self.assertEqual(source.is_active, True)
-        self.assertEqual(source.is_public, True)
+        self.assertEqual(source.is_active, is_active)
+        self.assertEqual(source.is_public, is_public)
         self.assertEqual(source.create, True)
 
     def assert_successful_add_production_location_without_geocode_result(